requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
bokeh==3.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
crispy-bootstrap5==0.7
cryptography==40.0.1
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.18
django-ajax-datatable==4.4.5
django-chamber==0.6.17
django-crispy-forms==2.0
django-datatable==0.3.1
django-datatables-view==1.20.0
django-import-export==3.2.0
django-jazzmin==2.6.0
django-pandas==0.6.6
django-pdfkit==0.3.1
django-pynotify==0.5.5
django-registration-redux==2.12
et-xmlfile==1.1.0
filemagic==1.6
frozenlist==1.3.3
html2pdf==0.1
html5lib==1.1
idna==3.4
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.2
multidict==6.0.4
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
pdfkit==0.6.0
Pillow==9.4.0
plotly==5.13.1
prompt-toolkit==3.0.38
pycparser==2.21
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pypdf==3.7.0
pypng==0.20220715.0
PyPrind==2.11.3
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
svglib==1.5.1
tablib==3.4.0
tenacity==8.2.2
tinycss2==1.2.1
tornado==6.2
twilio==8.0.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
Unidecode==1.3.6
uritools==4.0.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.9
xlrd==2.0.1
xlwt==1.3.0
xyzservices==2023.2.0
yarl==1.8.2
